// src/redux/action-types/AttackProcess.types.ts
export enum AttackProcessActionTypes {
  TOOGLE_ASK_SCAN = "TOOGLE_ASK_SCAN",
  START_ATTACK_PENDING = "START_ATTACK_PENDING",
  START_ATTACK_SUCCESS = "START_ATTACK_SUCCESS",
  START_ATTACK_FAILED = "START_ATTACK_FAILED",
  START_SCANING = "START_SCANING",
  SCANING_SUCCESS = "SCANING_SUCCESS",
  SCANING_FAILED = "SCANING_FAILED",
  ADD_DETAIL_PROCESS = "ADD_DETAIL_PROCESS",
}